requirement.txt
beautifulsoup4
sentence-transformers
playwright
python-dotenv
pinecone
streamlit
google-genai
"fastapi[standard]"